backend/pyproject.toml
[tool.poetry]
name = "componentsdb"
version = "0.1.0"
description = "A web application implementing an electronic components database"
authors = ["Rich Wareham <rich.componentsdb@richwareham.com>"]
readme = "../README.md"
packages = [{include = "componentsdb"}]

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.31"}
uvicorn = "^0.30.1"
structlog = "^24.4.0"
asyncpg = "^0.29.0"
faker = "^26.0.0"
strawberry-graphql = {extras = ["cli", "fastapi"], version = "^0.236.0"}
jwcrypto = "^1.5.6"
validators = "^0.33.0"
requests-cache = "^1.2.1"
fastapi = "^0.111.1"
pydantic = "^2.8.2"
pydantic-settings = "^2.3.4"
python-multipart = "^0.0.9"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
alembic = "^1.13.2"
poethepoet = "^0.27.0"
pytest-cov = "^5.0.0"
coverage = "^7.6.0"
better-exceptions = "^0.3.3"
pytest-docker-tools = "^3.1.3"
pytest-asyncio = "^0.23.8"
faker = "^26.0.0"
responses = "^0.25.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks."test"]
help = "Run the pytest test suite"
cmd = "pytest"

[tool.pytest.ini_options]
addopts = "--cov --cov-report html --cov-report term"
filterwarnings = [
  # From strawberry GraphQL library
 "ignore:'typing.ByteString' is deprecated:DeprecationWarning",
]

[tool.coverage.run]
branch = true
source = ["componentsdb"]

[tool.black]
line-length = 99

[tool.isort]
profile = "black"

[tool.mypy]
ignore_missing_imports = true
plugins = ["strawberry.ext.mypy_plugin"]

[tool.datamodel-codegen]
input = "../openapi/spec.yaml"
input-file-type = "openapi"
output = "componentsdb/api/models.py"
output-model-type = "pydantic_v2.BaseModel"
use-annotated = true
field-constraints = true
snake-case-field = true
target-python-version = "3.12"
use-double-quotes = true
disable-timestamp = true
